% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fNMLComplexity.R
\name{computeComplexityOrient}
\alias{computeComplexityOrient}
\title{Compute complexity NML non Markov equivalent between two variables with
conditioning set}
\usage{
computeComplexityOrient(X, Y, df_conditioning = NULL)
}
\arguments{
\item{X}{[vector]
A vector that contains the observational data of the first variable.}

\item{Y}{[vector]
A vector that contains the observational data of the second variable.}

\item{df_conditioning}{[data frame]
The data frame of the observations of the conditioning variables. Parents of
Y node}
}
\value{
Numeric value of the complexity NML term
}
\description{
The complexity term is calculated knowing that X is the parent of Y
such that X -> Y. So the conditioning set corresponds to the parents of node
Y
}
\examples{
N <- 1000
X <- sample(x = c(1, 2, 3, 4, 5), size = N, replace = TRUE)
Y <- sample(x = c(1, 2, 3), size = N, replace = TRUE)
Z <- sample(x = c(5, 6), size = N, replace = TRUE)
res <- computeComplexityOrient(as.factor(X), as.factor(Y), df_conditioning = as.factor(Z))
message("k(X;Y|Z) = ", res)
}
