% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoreEdge.R
\name{computeEdgeScore}
\alias{computeEdgeScore}
\title{Compute edge score (mutual information) with conditioning set}
\usage{
computeEdgeScore(X, Y, df_conditioning = NULL)
}
\arguments{
\item{X}{[vector]
A vector that contains the observational data of the first variable.}

\item{Y}{[vector]
A vector that contains the observational data of the second variable.}

\item{df_conditioning}{[data frame]
The data frame of the observations of the conditioning variables.}
}
\value{
A list that contains :
\itemize{
\item info: The estimation of (conditional) mutual information without the
complexity cost.
\item infok: The estimation of (conditional) mutual information with the
complexity cost (\eqn{Ik = -I + cplx}).}
}
\description{
Compute edge score (mutual information) with conditioning set and Markov
equivalent (symmetric) complexity NML term
}
\examples{
N <- 1000
X <- sample(x = c(1, 2, 3, 4, 5), size = N, replace = TRUE)
Y <- sample(x = c(1, 2, 3), size = N, replace = TRUE)
res <- computeEdgeScore(as.factor(X), as.factor(Y))
message("I(X;Y) = ", res$info)
message("Ik(X;Y) = ", res$infok)
Z <- sample(x = c(5, 6), size = N, replace = TRUE)
res <- computeEdgeScore(as.factor(X), as.factor(Y), df_conditioning = as.factor(Z))
message("I(X;Y|Z) = ", res$info)
message("Ik(X;Y|Z) = ", res$infok)
}
