% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covidCases}
\alias{covidCases}
\title{Covid cases}
\format{
A data.frame object.
}
\usage{
data(covidCases)
}
\description{
Demo dataset of chronological series to be used in temporal mode of miic.
Evolution of Covid cases on a subset of EU countries from 12/31/2019 to 06/18/2020.
Source of the data : European Centre for Disease Prevention and Control.
}
\references{
ECDC (\href{https://www.ecdc.europa.eu/en/publications-data/download-todays-data-geographic-distribution-covid-19-cases-worldwide}{ECDC link})
}
\keyword{data}
\keyword{datasets}
