% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hematoData}
\alias{hematoData}
\title{Early blood development: single cell binary gene expression data}
\format{
A data.frame object.
}
\usage{
data(hematoData)
}
\description{
Binarized expression data of 33 transcription factors involved
in early differentiation of primitive erythroid and endothelial
cells (3934 cells).
}
\references{
Moignard et al. (2015) Nat Biotechnol 33(3):269-76
(\doi{10.1038/nbt.3154}).
}
\keyword{data}
\keyword{datasets}
