% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactmiic.utils.R
\name{interactmiic.CCClinks}
\alias{interactmiic.CCClinks}
\title{This function identifies cell-cell communication links to include in the 
interactMIIC network using the LIANA package.}
\usage{
interactmiic.CCClinks(
  seurat_object,
  species = "human",
  assay_name = "RNA",
  CCC_method = "cellphonedb",
  n_CCClinks = 30,
  senders,
  receivers,
  interact_ident
)
}
\arguments{
\item{seurat_object}{A Seurat object containing single-cell data.}

\item{species}{A character string indicating the species. Either "human" 
(default) or "mouse".}

\item{assay_name}{A character string specifying the assay name in the 
Seurat object. Default is "RNA".}

\item{CCC_method}{A character string specifying the CCC method to use, see 
show_methods() of LIANA}

\item{n_CCClinks}{A numeric specifiying how many L-R to keep. Default is 60.}

\item{senders}{A non-empty character vector of sender cell types.}

\item{receivers}{A non-empty character vector of receiver cell types.}

\item{interact_ident}{A character string specifying the identity column in 
the Seurat object where senders and receivers can be found.}
}
\value{
A data frame containing the top cell-cell communication links.
}
\description{
This function generates an LIANA output of senders and receivers
significant interactions.
}
\examples{
\dontrun{
result <- interactmiic.CCClinks(seurat_obj, species = "human", senders = c("T_cells"), 
receivers = c("B_cells"), interact_ident = "celltype")
}

}
