% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactmiic.utils.R
\name{interactmiic.layout}
\alias{interactmiic.layout}
\title{Create a layout file for interactMIIC}
\usage{
interactmiic.layout(
  state_order,
  labelSize = 16,
  threshold = 0,
  edgeLength = 120,
  network_height = NULL,
  CCC_sizefactor = 1.3,
  node_spacing = 100
)
}
\arguments{
\item{state_order}{A data frame containing the state order information.
output of interactmiic.state_order()}

\item{labelSize}{The size of the labels in the layout. Default is 16.}

\item{threshold}{The threshold for log confidence on the edges. Default is 0.}

\item{edgeLength}{The default edge length for the network. Default is 120.}

\item{network_height}{The height of the  network grids for gene senders and
receivers in terms of number of nodes. Default is automatically calculated 
from the number of genes in the network.}

\item{CCC_sizefactor}{The multiplier factor for the ligands and receptors
grids height (network_height*CCC_sizefactor). Default is 1.3.}

\item{node_spacing}{The spacing between nodes in the layout. Default is 100.}
}
\value{
A json file with positions for each node of the network. See how to
include it in your interactMIIC job on the MIIC webserver.
}
\description{
This function generates an automatic json file to input into
your interactMIIC job so your network benefits from a clear interaction
display. This initial layout can be modified with your mouse cursor or
with the "Relax" button on the MIIC server for a more organic display.
}
