% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactmiic.utils.R
\name{interactmiic.MIselection}
\alias{interactmiic.MIselection}
\title{Select informative genes for the network reconstruction}
\usage{
interactmiic.MIselection(
  seurat_object,
  assay_name = "RNA",
  interact_ident,
  oneinteract,
  goi,
  metadata_list,
  n_genes = min(length(goi) * 5, 100),
  plot = FALSE,
  save = FALSE,
  output_dir = "MI_tables"
)
}
\arguments{
\item{seurat_object}{[a Seurat object]
A Single-Cell transcriptomics Seurat object}

\item{assay_name}{[a string]. Gives the name of the assay to take
the transcriptomics raw counts from (usually 'RNA')}

\item{interact_ident}{[a string]. Gives the name of the metadata
containing the celltypes population}

\item{oneinteract}{[a string]. Either the senders tag or the receivers tag}

\item{goi}{[a vector] A list of genes of interest in your dataset.
Must be of length <15. These genes will be individually used as pivot
 variable to look for mutual information in your dataset.}

\item{metadata_list}{[a vector] A list of metadata names of interest in
your dataset. Must be of length <15. These metadata will be individually
used as pivot variable to look for mutual information in your dataset.}

\item{n_genes}{[a numeric] The number of genes to keep after pairwise
mutual information ranking. Default is 30}

\item{plot}{[a boolean] A boolean to specify if you want a heatmap plot
of the highest mutual information found with the features of interest,
default is FALSE. Will be saved in output_dir}

\item{save}{[a boolean] A boolean to specify if you want to save the full
mutual information table, default is FALSE. Will be saved in output_dir}

\item{output_dir}{[a string] A string to specify an output directory to
export optional outputs to, default is "MI_tables"}
}
\value{
A vector of genes of length n_genes sharing the most mutual
information with the given features of interest
}
\description{
This function offers an unsupervized feature selection tool
based on fast pairwise mutual information computation. If the user do not
gives genes_senders/genes_receivers, this function will select relevant
genes for the interactMIIC network
}
