% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactmiic.utils.R
\name{interactmiic.state_order}
\alias{interactmiic.state_order}
\title{Create a state order file for interactMIIC}
\usage{
interactmiic.state_order(
  mosaic_data_table,
  metadata_senders,
  metadata_receivers,
  genes_senders,
  ligands,
  receptors,
  genes_receivers
)
}
\arguments{
\item{mosaic_data_table}{[a data frame]
An output of interactmiic.mosaic()}

\item{metadata_senders}{[a vector] A named list of metadata (strings) with
levels as items. See exemple}

\item{metadata_receivers}{[a vector] A named list of metadata (strings) with
levels as items. See exemple}

\item{genes_senders}{[a vector] A list of selected genes (strings) for the
senders cells. Overlap with genes_receivers is not an issue}

\item{ligands}{[a vector] A list of selected CCC genes (strings) for the
senders cells. Overlap with genes_senders is not an issue}

\item{receptors}{[a vector] A list of selected CCC genes (strings) for the
receivers cells. Overlap with genes_receivers is not an issue}

\item{genes_receivers}{[a vector] A list of selected genes (strings) for the
receivers cells. Overlap with genes_senders is not an issue}
}
\value{
A state order that can be used for interactMIIC
network reconstruction.
}
\description{
The state order is an mandatory file that allows you to input
optional information for the computation (such as contextual variables,
types of variables (otherwise detected automatically)... ) and information
for the display (groups of nodes, levels ordering of categorical data ...).
}
\examples{
\donttest{
metadata <- list(
 treatment = c("Control", "Treated"),
 another_meta = c("Level1", "Level2", "Level3")
)
}

}
