% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactmiic.wrapper.R
\name{interactmiic.wrapper}
\alias{interactmiic.wrapper}
\title{Run the preprocessing of an Seurat object for interactMIIC}
\usage{
interactmiic.wrapper(
  seurat_object,
  assay_name = "RNA",
  interact_ident = NULL,
  senders_name = NULL,
  receivers_name = NULL,
  goi_senders = NULL,
  goi_receivers = NULL,
  genes_senders = NULL,
  genes_receivers = NULL,
  species = "human",
  do_CCC = TRUE,
  CCC_method = "cellphonedb",
  n_CCClinks = 20,
  ligands = NULL,
  receptors = NULL,
  metadata_senders = NULL,
  metadata_receivers = NULL,
  wd_path = getwd()
)
}
\arguments{
\item{seurat_object}{[a Seurat object]
A Single-Cell transcriptomics Seurat object}

\item{assay_name}{[a string]. Gives the name of the assay to take
the transcriptomics raw counts from (usually 'RNA')}

\item{interact_ident}{[a string]. Gives the name of the metadata
containing the celltypes population}

\item{senders_name}{[a string]. Gives the name of the senders}

\item{receivers_name}{[a string]. Gives the name of the receivers}

\item{genes_senders}{[a vector] A list of selected genes (strings) for the
senders cells. Overlap with genes_receivers is not an issue}

\item{genes_receivers}{[a vector] A list of selected genes (strings) for the
receivers cells. Overlap with genes_senders is not an issue}

\item{metadata_senders}{[a vector] A list of metadata (strings) that will
be added to the genes network. They can be relevant to senders
and/or receivers cells.}

\item{metadata_receivers}{[a vector] A list of metadata (strings) that will
be added to the genes network. They can be relevant to senders
and/or receivers cells.}
}
\value{
A mosaic datatable that can be used for interactMIIC
network reconstruction.
}
\description{
Function that takes an seurat_object raw counts,
 senders and receivers information metadata and create a mosaic
 datable that contains first senders network
 variables then receivers network variables
}
