% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_efn.R
\name{miic.evaluate.effn}
\alias{miic.evaluate.effn}
\title{Evaluate the effective number of samples}
\usage{
miic.evaluate.effn(input_data = NULL, plot = T)
}
\arguments{
\item{input_data}{[a data frame]
A data frame that contains the observational data. Each
column corresponds to one variable and each row is a sample that gives the
values for all the observed variables. The column names correspond to the
names of the observed variables. Data must be discrete like.}

\item{plot}{[a boolean value] if the autocorrelation plot has to be done. It will be performed only if all values of the correlation vector are positive.}
}
\value{
A list containing the autocorrelation decay, the effective number of samples, and the result of an exponentiality test with alpha = 0.05
}
\description{
This function evaluates the effective number of samples in a dataset.
}
