% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miic.export.R
\name{miic.export}
\alias{miic.export}
\title{Export miic result to different plotting methods}
\usage{
miic.export(miic.res, method = NULL)
}
\arguments{
\item{miic.res}{[a miic graph object]
The graph object returned by the miic execution.}

\item{method}{A string representing the plotting method. Currently only "igraph" is supported}
}
\value{
A graph object adapted to the method.
}
\description{
This function returns an object built from the results
of the miic main function that is ready to be fed to different
plotting methods.
}
\details{
See the details of specific function for each method.
}
\examples{
\dontrun{
library(miic)

# Using igraph
library(igraph)
g = miic.export(miic.res, "igraph")
plot(g) # Default visualisation, calls igraph::plot.igraph()

}

}
\seealso{
\code{\link{getIgraph}} for returning an igraph object built from the results.
}
