% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.cytoscape.R
\name{miic.write.network.cytoscape}
\alias{miic.write.network.cytoscape}
\title{GraphML converting function for miic graph}
\usage{
miic.write.network.cytoscape(g, file, layout = NULL)
}
\arguments{
\item{g}{The graph object returned by \link[=miic]{miic}.}

\item{file}{A string. Path to the output file containing file name without
extension (.graphml will be appended).}

\item{layout}{An optional data frame of 2 (or 3) columns containing the
coordinate \code{x} and \code{y} for each node. The optional first column can contain
node names. If node names is not given, the order of the input file will be
assigned to the list of positions.}
}
\description{
Convert miic graph to \href{http://graphml.graphdrawing.org/}{GraphML format}.
}
