% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.style.R
\name{miic.write.style.cytoscape}
\alias{miic.write.style.cytoscape}
\title{Style writing function for the miic network}
\usage{
miic.write.style.cytoscape(file)
}
\arguments{
\item{file}{[a string] The file path of the output file (containing the file name without extension).}
}
\description{
This function writes the  miic style for a correct visualization using the cytoscape tool (http://www.cytoscape.org/).
}
\details{
The style is written in the xml file format.
}
