% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ohno_stateOrder}
\alias{ohno_stateOrder}
\title{Tetraploidization in vertebrate evolution}
\format{
A data.frame object.
}
\usage{
data(ohno_stateOrder)
}
\description{
20,415 protein-coding genes in the human genome from Ensembl (v70) and information on the
retention of duplicates originating either from the two whole genome duplications at
the onset of vertebrates (‘ohnolog’) or from subsequent small scale duplications (‘SSD’)
as well as copy number variants (‘CNV’), category order.
}
\references{
Verny et al., PLoS Comp. Bio. 2017.
}
\keyword{data}
\keyword{datasets}
