% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmiic.plot.R
\name{plot.tmiic}
\alias{plot.tmiic}
\title{Basic plot function of a tmiic network inference result}
\usage{
\method{plot}{tmiic}(tmiic.res, method = "igraph", ...)
}
\arguments{
\item{tmiic.res}{[a tmiic graph object]
The graph object returned by \code{\link{miic}} in temporal mode, 
eventually flatten}

\item{method}{A string representing the plotting method. Default to "igraph".
Currently only "igraph" is supported.}

\item{\dots}{Additional plotting parameters. See the corresponding plot function
for the complete list.

For igraph, see \code{\link[igraph]{igraph.plotting}}.}
}
\description{
This function calls \code{\link{tmiic.export}} to build a
plottable object from the result returned by \code{\link{miic}} in 
temporal mode and plot it.
}
\details{
See the documentation of \code{\link{tmiic.export}} for further
details.
}
\examples{
library(miic)

#' # EXAMPLE COVID CASES (timeseries demo)
data(covidCases)
# execute MIIC (reconstruct graph in temporal mode)
tmiic.res <- miic(input_data = covidCases, tau = 2, movavg = 14)

# plot temporal graph
if(require(igraph)) {
 plot(tmiic.res)
}

# to plot a condensed graph
flatten.res <- tmiic.flatten_network(tmiic.res)
if(require(igraph)) {
 plot(flatten.res)
}
}
\seealso{
\code{\link{tmiic.export}} for generic exports,
\code{\link{tmiic.getIgraph}} for igraph export,
\code{\link[igraph]{igraph.plotting}}
}
