% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fNMLComplexity.R
\name{regret}
\alias{regret}
\title{Compute regret term}
\usage{
regret(N, r, data)
}
\arguments{
\item{N}{[an integer]
Length of the categotical data}

\item{r}{[an integer]
Number of different categories (variable level) data}

\item{data}{[a data frame]
A data frame of the same length N as the variable studied}
}
\value{
Numeric value of the regret term
}
\description{
The logarithm of the denominator of Normalized Maximum Likelihood (NML) is
often called the regret
}
\examples{
library(miic)
N <- 1000
r <- 45
X <- sample(x = c(1, 2, 3, 4, 5), size = N, replace = TRUE)
res <- regret(N, r, data.frame(X))
message("reg(N, r) = ", res)

}
\references{
\itemize{
\item Silander et al., \emph{AISTATS 2018.}
  https://proceedings.mlr.press/v84/silander18a/silander18a.pdf}
}
