% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectFeaturesPath.R
\name{selectFeaturesPath}
\alias{selectFeaturesPath}
\title{Features selection on the path between variables of interest}
\usage{
selectFeaturesPath(
  input_data,
  var_of_interest_names_side1 = NULL,
  var_of_interest_values_side1 = NULL,
  var_of_interest_names_side2 = NULL,
  var_of_interest_values_side2 = NULL,
  n_selected = 1,
  n_alternate = 2,
  max_features = 250,
  max_depth = 4,
  precomputed_values = NULL,
  corrected = T,
  skip_cheks = F,
  n_threads = 1,
  verbose = 2,
  plot = F,
  ...
)
}
\arguments{
\item{input_data}{[a data frame or a matrix, required]

Expected layout is samples as rows and variables as columns. Column names
must contain the names of the variables.}

\item{var_of_interest_names_side1}{[a string or vector of strings, optional,
NULL by default]

For the variable(s) of interest that are part of the \emph{input_data}
on the first side of the path, you should supply their names here.}

\item{var_of_interest_values_side1}{[a data frame, optional, NULL by default]

For the variables of interest that are not in \emph{input_data}
and on the first side of the path, a data frame can be supplied.
The column names are the names of the variables of interest
and rows are the samples ordered in the same way as the \emph{input_data}.
Typically, such variables are metadata associated to samples but not
stored in \emph{input_data}, e.g. a "Treatment" vs "Control" variable
in an experiment and a count matrix with the expression of genes
in \emph{input_data}.}

\item{var_of_interest_names_side2}{[a string or vector of strings, optional,
NULL by default]

Same as \emph{var_of_interest_names_side1} for the second side of the path.}

\item{var_of_interest_values_side2}{[a data frame, optional, NULL by default]

Same as \emph{var_of_interest_values_side1} for the second side of the path.}

\item{n_selected}{[a positive integer, optional, 1 by default]

The number of features selected at each step, for each path explored
and at each iteration of the recursion. Increasing this value
will identify more variables as possible features at each step
with the consequence to reach quicker a target \emph{max_features}
while increasing the false positives.}

\item{n_alternate}{[a positive integer, optional, 2 by default]

The number of alternate paths to explore at each step of the recursion.
Decreasing this value speeds up the process while reducing the number
of possible paths discovered. Increasing it has opposite effect,
more features from alternate possible paths at the cost of an increased
processing time.}

\item{max_features}{[a positive integer, optional, 250 by default]

The maximum total number of features to select. Set by default to a large
number of features as the goal is to not reach this maximal number while
finding all variables on the path between the variables of interest.
Consider to tune the \emph{max_depth} and \emph{n_alternate} parameters
to identify more or less features.}

\item{max_depth}{[a positive integer, optional, 4 by default]

The maximum depth of the recursion.
Decreasing this value speeds up the process while reducing the granularity
of the path discovery. Increasing it has opposite effect, more features
can be collected along the path at the cost of an increased processing time.}

\item{precomputed_values}{[a list, optional, NULL by default]

If path features selection have been previously performed on the same
conditions (same variables of interest, same \emph{input_data},
same \emph{corrected} parameter),
Supplying the result of a previous run speeds up the process
as the existing computed values will be reused.
The expected format of the list is the same as the function returned value,
with at least an item "mis" containing a sub-list of \emph{mis} matrices
and an "scores" item containing a sub-list of \emph{scores} data frames.}

\item{corrected}{[a boolean, optional, TRUE by default]

When set to TRUE, the mutual information and 3 points information
are corrected by subtracting a complexity term
(computed with the Normalized Maximum Likelihood).
For dataset having very few samples, the complexity term can have
a disproportionate impact. Setting \emph{corrected} to FALSE switches
to the use of non corrected mutual information.}

\item{skip_cheks}{[a boolean, optional, FALSE by default]

Before computing MI between the variable of interest and the features,
\emph{input_data} is checked to filter out constant features and rows full
of NAs. When the \emph{input_data} does not need such filtering,
these checks can be skipped to speed up the process.}

\item{n_threads}{[a positive integer, optional, 1 by default]

When set greater than 1, \emph{n_threads} parallel threads are used for
computation. Make sure your compiler is compatible with openmp
if you wish to use multithreading.}

\item{verbose}{[an integer, optional, 2 by default]

Level of verbosity: 0=no display, 1=summary, 2=progress per couple of
variable of interest, 3=include more detail on the progress during
the recursion.}

\item{plot}{[a boolean, optional, FALSE by default]

If set to TRUE, a plot with the top features is generated
(requires `ggplot2`).

Please note on the plot rendering that positions are indicative
as they are only based on the ratio of MI shared between each feature
and the variables of interest of each side.}

\item{...}{If plotting is requested, extra parameters can be used to customize the plot
rendering:

\itemize{
\item \emph{x_lab}: a string, optional, "Features for
  variables_of_interest_side1-variables_of_interest_side2" by default,
  the X axis label.
\item \emph{depth_plot}: an integer between 1 and 10, optional, 4 by default,
  maximal depth used for the plot.
\item \emph{annotate}: a boolean, optional, TRUE by default,
  When activated, displays the maximal depth of the recursion
  and the number of features not displayed on the plot.
\item \emph{font_size}: an integer, optional, 11 by default, the font size.
}}
}
\value{
A named list with five items:

\itemize{
\item \emph{features}: a vector with the features selected.
\item \emph{mis}: a list of matrices with the MIs between the variables in
  \emph{input_data} (as rows) and the variable(s) of interest (as columns).
  Row and column names are sorted alphabetically.\cr
  The MIs can be corrected or not, depending on lthe \emph{corrected}
  parameters. If pre-computed values were supplied, new values computed
  are added to the existing matrices.
\item \emph{couples}: a data frames with, at each depth, the couples
  of variables used with their MIs and features selected.
  Please note that couples already encountered are not included
  as no recursion is performed on duplicates.
\item \emph{scores}: a list of data frames with, for all pairs of variables
  passing the dpi test, the information about the contionning set used,
  the MI values, the DPI test, the 3 points information and the scores.
\item \emph{plot}: when the \emph{plot} parameter is turned to TRUE,
  a plot with the features selected, NULL otherwise.
  Please note on the plot rendering that positions are indicative
  as they are only based on the ratio of MI shared between each feature
  and the variables of interest of each side.
}
}
\description{
Select the variables that are the most likely to be in the path
between two set of variable(s) of interest.
Variables selection is performed using a score combining Data Processing
inequality (DPI) and 3 points information (see Affeldt 2015).\cr
The search is recursive by dichotomy: in the first round, a set of features
on the path between the main variables of interest is selected.
From then, the path is split in two parts:
\itemize{
  \item variables of interest of side 1 + set of features selected
  \item set of features selected + variables of interest of side 2
  }
Each part of the path is investigated with the same principle:
finding a set of features in the path subsection, splitting the path in two,
and so on.
The recursion ends when no feature can be found
or when the maximum level of recursion is reached.
}
\examples{
library(miic)

\donttest{
# Features selection on the path between an external metadata "Ploidy"
# (simulation by extracting "Ploidy" out of the dataset)
# and the gene expression of TP53
df_external_meta <- data.frame ("Ploidy"=cosmicCancer$Ploidy)
df_data <- cosmicCancer[ , ! (colnames(cosmicCancer) == "Ploidy") ]
ret <- selectFeaturesPath (input_data=df_data,
                           var_of_interest_values_side1=df_external_meta,
                           var_of_interest_names_side2="TP53")
message ("Features selected: ", paste (ret$features, collapse=", ") )

# Same features selection with reuse of the previous run above and plot
ret <- selectFeaturesPath (input_data=df_data,
                           var_of_interest_values_side1=df_external_meta,
                           var_of_interest_names_side2="TP53",
                           precomputed_values=ret, plot=TRUE)
print (ret$plot)

# Features selection using multiple variables on each side
# (limiting the recursion depth to speed up the example)
# and selecting two features per step
ret <- selectFeaturesPath (input_data=cosmicCancer,
                           var_of_interest_names_side1=c("TP53", "Ploidy"),
                           var_of_interest_names_side2=c("FOXM1", "AURKA"),
                           max_depth=2, n_selected=2)
message ("Features selected: ", paste (ret$features, collapse=", ") )

# Similar features selection with plotting using a customized rendering
# (reusing the previous run above and limiting the recursion depth
# to speed up the example)
ret <- selectFeaturesPath (cosmicCancer,
                           var_of_interest_names_side1=c("TP53", "Ploidy"),
                           var_of_interest_names_side2=c("FOXM1", "AURKA"),
                           max_depth=2, precomputed_values=ret,
                           plot=TRUE, font_size=12,
                           x_lab="My features selection on CosmicCancer")
print (ret$plot)
}

}
\references{
\itemize{
\item Affeldt \emph{et al.}, UAI 2015, \href{https://auai.org/uai2015/proceedings/papers/293.pdf}{Robust Reconstruction of Causal Graphical Models based on Conditional 2-point and 3-point Information}
}
}
