% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmiic.datasets.R
\name{tmiic.call_funct}
\alias{tmiic.call_funct}
\title{tmiic.call_funct}
\usage{
tmiic.call_funct(funct, x)
}
\arguments{
\item{funct}{[a function] The function to use. 
Typically, the function will be \emph{tmiic.f1}, \emph{tmiic.f2} or 
\emph{tmiic.f3} but any function accepting a float as parameter and 
returning a float can be used.}
}
\value{
the number computed by the funct function
}
\description{
Utility function to call a function supplied as parameter.
}
