% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmiic.scores.R
\name{tmiic.compute_list_scores}
\alias{tmiic.compute_list_scores}
\title{tmiic.compute_list_scores}
\usage{
tmiic.compute_list_scores(list_res, df_true_edges, list_nodes, tau)
}
\arguments{
\item{list_res}{[a list] the list of tmiic's results. It can be a list of
full results or only summary}

\item{df_true_edges}{[a dataframe] The list of the true edges}

\item{list_nodes}{[a list] The list of nodes in the network}

\item{tau}{[an integer] The number of timesteps back in time used to 
construct the temporal graphs}
}
\value{
[a list] the returned list contains two dataframes containing 
both the list of scores for each tmiic result. The first dataframe is for
unoriented edges and the second about oriented ones.
}
\description{
Compute scores on a list of tmiic results
}
