% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmiic.scores.R
\name{tmiic.compute_scores}
\alias{tmiic.compute_scores}
\title{tmiic.compute_scores}
\usage{
tmiic.compute_scores(df_edges, df_true_edges, list_nodes, tau)
}
\arguments{
\item{df_edges}{[a dataframe] The list of edges computed by tmiic
(the summary dataframe of tmiic's return)}

\item{df_true_edges}{[a dataframe]  The list of true edges}

\item{list_nodes}{[a list] The list of nodes in the network}

\item{tau}{[an integer] The number of timesteps back in time used to 
construct the temporal graph}
}
\value{
[a list] The returned list contains two data frames containing 
both the confusion matrix, the precision, recall and f-score. The first 
dataframe is for unoriented edges and the second about oriented ones.
}
\description{
Compute confusion matrix, precision, recall and f-score on the 
reconstructed graph. Values are computed both for unoriented and 
oriendted graphs
}
