% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmiic.datasets.R
\name{tmiic.f1}
\alias{tmiic.f1}
\title{tmiic.f1}
\usage{
tmiic.f1(x)
}
\arguments{
\item{x}{[a number]}
}
\value{
the number received as argument
}
\description{
basic linear function used for samples generation f(x) = x
}
