% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmiic.datasets.R
\name{tmiic.f3}
\alias{tmiic.f3}
\title{tmiic.f3}
\usage{
tmiic.f3(x)
}
\arguments{
\item{x}{[a number]}
}
\value{
the number computed by f(x)
}
\description{
non linear function used for samples generation 
\eqn{f(x) = ( 1 - 4 * x^3 * exp[-(x^2)/2] ) * x}
}
