% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmiic.datasets.R
\name{tmiic.generate_predefined_dataset}
\alias{tmiic.generate_predefined_dataset}
\title{tmiic.generate_predefined_dataset}
\usage{
tmiic.generate_predefined_dataset(
  model_idx,
  funct,
  n_timeseries,
  n_timesteps,
  seed = NULL
)
}
\arguments{
\item{model_idx}{[an integer betwen 0 and 7] The model for the data 
generation:
\itemize{
\item Model 0 is for debug (intialazing nodes with easily identiable values)
\item Model 1 uses only t-1 information. 
\item Models 2 and 3 use both t-1 and t-2.
\item Model 4 uses t-1, t-2 and contemporaneous edge
\item Model 5 isolated temporal edges oriented using time
\item Model 6 contemporaneous edge oriented using the 4 points condition
\item Model 7 latent node unveiled by edge between X2 and X4
}}

\item{funct}{[a function] the function used for the generation.\cr
It can be any function computing a float from a float or one of the
predefined ones:
\itemize{
\item \emph{tmiic.f1}: \eqn{f(x) = x}
\item \emph{tmiic.f2}: \eqn{f(x) = ( 1 - 4 * exp[-(x^2)/2] ) * x}
\item \emph{tmiic.f3}: \eqn{f(x) = ( 1 - 4 * x^3 * exp[-(x^2)/2] ) * x}
}}

\item{n_timeseries}{[an integer] The number of samples to generate}

\item{n_timesteps}{[an integer] The number of timesteps of the time series 
generated}

\item{seed}{[an integer] Optional, NULL by default. The seed to use when 
generating the dataset}
}
\value{
a list with two entries :
\itemize{
\item a dataframe of \emph{n_timeseries} * \emph{n_timesteps} rows 
and \emph{n_nodes} columns
\item a dataframe with the true edes of the network
}
}
\description{
Generate n_timeseries samples for x nodes over n_timesteps timesteps
using a predefined model
}
