% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmiic.plot.R
\name{tmiic.plot}
\alias{tmiic.plot}
\title{Igraph plotting function for tmiic (temporal miic)}
\usage{
tmiic.plot(
  g,
  method = "log_confidence",
  igraph_layout = NULL,
  user_layout = NULL,
  miic_defaults = TRUE,
  filename = NULL,
  file_figsize = NULL,
  font_family = NULL,
  title = NULL,
  title_cex = 1.5,
  draw_legend = TRUE,
  graph_margin = NULL,
  nodes_label_dist = NULL,
  nodes_label_degree = NULL,
  nodes_label_cex = NULL,
  nodes_shape = NULL,
  nodes_colors = NULL,
  nodes_sizes = NULL,
  edges_label_cex = NULL,
  edges_width = NULL,
  edges_curved = NULL,
  edges_arrow_size = NULL,
  edges_arrow_width = NULL,
  verbose = FALSE
)
}
\arguments{
\item{g}{[a miic graph object]
The graph object returned by the miic execution.}

\item{method}{[a string; \emph{c("pcor", "log_confidence")}]
Optional, "log_confidence" by default. The column used to plot the 
strength of the edges.}

\item{igraph_layout}{[an igraph layout object]
Optional, \emph{layout_with_kk} by default. When set, it is used to plot the network. 
See the igraph manual for more information.}

\item{user_layout}{[a data frame]
Optional, NULL by default. A data frame reporting the position of nodes. 
Each line corresponds to the \emph{(x,y)} coordinates of each vertex.
When the data frame has two columns, the first one is assocated with
\emph{x} and the second to \emph{y}. When the data frame has more than 
two columns, columns two and three are extracted and mapped, 
respectively, to \emph{x} and \emph{y} positions.}

\item{miic_defaults}{[a boolean] Optional, TRUE by default. 
When TRUE, several graphic parameters are intialised with default values
suited for most of miic results:
\itemize{
\item For nodes, font family: "Helvetica", color: "lightblue", size: 10, 
label.cex: 0.6
\item For edges, arrow.size: 0.2, arrow.width: 3, width: 3, curved: FALSE
}
Note that if the user supplies specific values for these parameters
(ie a specific node size), the user choices will overwrite the miic 
defaults.}

\item{filename}{[a string] Optional, NULL by default. 
If supplied, plot will be saved here.}

\item{file_figsize}{[a list; \emph{c(length, height)}] Optional, NULL by default. 
When plots are drawn to a file, the size of the draw in pixels can be specified
by a couple of values. The first is the lentgh, the second the height.}

\item{font_family}{[a string] Optional, NULL by default. The font to use.\cr
Note that the font may not apply in all the graphical parts, depending
on how the plotted objects manage the font parameter.}

\item{title}{[a string] Optional, NULL by default. The title of the plot.}

\item{title_cex}{[a number] Optional, 1.5 by default. The size of the title.}

\item{draw_legend}{[a boolean] Optional, TRUE by default. 
When TRUE, a legend is drawn on the right side of the plot.}

\item{graph_margin}{[a list of floats: c(bottom,left,top,right)]
Optional, NULL by default. Margin applied around the graph. 
When NULL, the value is determined by the presence or absence of 
self loop(s) in the graph: 
\itemize{
\item no self loop = no margin 
\item at least one self loop = margin of 0.25 on the left and right sides 
}}

\item{nodes_label_dist}{[a number] Optional, NULL by default. 
The distance of the labels from the nodes.}

\item{nodes_label_degree}{[a number] Optional, NULL by default.
The degre where to draw the labels from the nodes.}

\item{nodes_label_cex}{[a number] Optional, NULL by default.
The size of the nodes labels.}

\item{nodes_shape}{[a string] Optional, NULL by default.
The shape of the nodes. See the igraph manual for more information.}

\item{nodes_colors}{[a string or a vector] Optional, NULL by default.
The colors of the nodes. See the igraph manual for more information.}

\item{nodes_sizes}{[a number or a vector] Optional, NULL by default.
The sizes of the nodes. See the igraph manual for more information.}

\item{edges_label_cex}{[a number] Optional, NULL by default.
The size of the edges labels.}

\item{edges_width}{[a number] Optional, NULL by default.
The width of the edges.}

\item{edges_curved}{[a boolean or a vector] Optional, NULL by default.
The curvatures to apply to the edges:\cr
\itemize{
\item when TRUE, edges are curved\cr
\item when FALSE, edges are straitgh\cr
\item when NULL, the function will choose the best curvature depending 
on the type of network: curved for a lagged graphs and straitgh for 
condensed graphs.
}
Note that mutiple edges between the same nodes will always be curved
whatever value has edges_curved. 
See the igraph manual for more information.}

\item{edges_arrow_size}{[a number] Optional, NULL by default.
The size of the edges arrows.}

\item{edges_arrow_width}{[a number] Optional, NULL by default.
The width of the edges arrows.}

\item{verbose}{[a boolean value] Optional, FALSE by default. If TRUE, 
debugging output is printed.}
}
\description{
This function plots the temporal network with the given layout
(if specified) using the igraph package.
}
\details{
The plot reports the partial correlation or the log_confidence 
as strength of the edges. The function process differently condensed and
lagged networks.

For lagged networks (the ones with nodes including '_lagX' at their names 
ends), the function will define an user layout if none is supplied 
and use curved edges.

For condensed networks (the network obtained after removing the '_lagX' 
at the end of the nodes names), the function will plot the lag 
information on the edges.
}
