% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmiic.datasets.R
\name{tmiic.plot_one_sample}
\alias{tmiic.plot_one_sample}
\title{tmiic.plot_one_sample}
\usage{
tmiic.plot_one_sample(data_tab, sample_idx, title = NULL, filename = NULL)
}
\arguments{
\item{data_tab}{[an array of dimensions n_samples * n_nodes * n_time]
The array containing the samples}

\item{sample_idx}{[an integer] The sample index to plot}

\item{title}{[a string] Optional, NULL by default. The title of the plot}

\item{filename}{[a string] Optional, NULL by default. If supplied, the plot
is saved in this file}
}
\value{
None
}
\description{
Plot one sample of a temporal serie
}
