% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmiic.datasets.R
\name{tmiic.plot_one_timeseries}
\alias{tmiic.plot_one_timeseries}
\title{tmiic.plot_one_timeseries}
\usage{
tmiic.plot_one_timeseries(
  df_data,
  timseseries_idx,
  title = NULL,
  filename = NULL
)
}
\arguments{
\item{df_data}{[an array of dimensions n_timeseries * n_nodes * n_timesteps]
The array containing the samples}

\item{timseseries_idx}{[an integer] The index of the time series to plot}

\item{title}{[a string] Optional, NULL by default. The title of the plot}

\item{filename}{[a string] Optional, NULL by default. If supplied, the plot
is saved in this file}
}
\value{
None
}
\description{
Plot one sample of a temporal serie
}
