% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmiic.scores.R
\name{tmiic.plot_scores}
\alias{tmiic.plot_scores}
\title{tmiic.plot_scores}
\usage{
tmiic.plot_scores(
  scores,
  title,
  list_labels,
  type = "PRFS",
  plot_fntn = FALSE,
  filename = NULL
)
}
\arguments{
\item{scores}{[a dataframe] A dataframe of scores. Expected columns are
tp, fp, tn, fn, precision, recall and fscore  whilst each row correspond 
to a run of tmiic.}

\item{title}{[a string] The tille of the plot}

\item{list_labels}{[a list] The list of labels on the X axis. Typically,
these labels are the settings used for the different runs.}

\item{type}{[a string ] Optional, \emph{"PRFS"} by default. 
Represents the type of the plot: \emph{"PRFS"} plots Precision, Recall 
and F-Score whilst the other type available \emph{"TPFP"} plots 
True/False positive rates.}

\item{plot_fntn}{[a boolean] Optional, FALSE by default. Applies only to
type \emph{"TPFP"} plot. If TRUE, displays also True and False negative
rates.}

\item{filename}{[a string] Optional, NULL by default. The file name where
to save the plot,}
}
\value{
None
}
\description{
plot evolution of precision, recall and f-score or TP, FP, FN, TN rates
}
