% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.style.R
\name{writeCytoscapeStyle}
\alias{writeCytoscapeStyle}
\title{Style writing function for the miic network}
\usage{
writeCytoscapeStyle(file)
}
\arguments{
\item{file}{[a string] The file path of the output file (containing the
file name without extension).}
}
\value{
None
}
\description{
This function writes the  miic style for a correct
visualization using the cytoscape tool (http://www.cytoscape.org/).
}
\details{
The style is written in the xml file format.
}
