% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretizeMDL.R
\name{discretizeMDL}
\alias{discretizeMDL}
\title{Discretize a real valued distribution}
\usage{
discretizeMDL(myDist = NULL, maxbins = 20)
}
\arguments{
\item{myDist}{[a vector]
A vector that contains the observational data.}

\item{maxbins}{[an int] The maximum number of bins to test for.}
}
\value{
A vector containing the cutpoints of the best discretization according to the MDL principle.
}
\description{
This function discretizes a distribution with MDL.
}
