% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretizeMutual.R
\name{discretizeMutual}
\alias{discretizeMutual}
\title{Discretize two distributions with}
\usage{
discretizeMutual(myDist1 = NULL, myDist2 = NULL, maxbins = NULL,
  plot = T)
}
\arguments{
\item{myDist1}{[a vector]
A vector that contains the observational data of the first variable.}

\item{myDist2}{[a vector]
A vector that contains the observational data of the second variable.}

\item{maxbins}{[an int] The maximum number of bins to test for.}
}
\value{
A list with the two vectors containing the cutpoints of the best discretization for both variables.
}
\description{
This function discretizes two distributions by maximizing their mutual information, with MDL.
}
