% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miic.export.R
\name{getIgraph}
\alias{getIgraph}
\title{Igraph plotting function for miic}
\usage{
getIgraph(miic.res)
}
\arguments{
\item{miic.res}{[a miic graph object]
The graph object returned by the miic execution.}
}
\value{
An igraph graph object.
}
\description{
This functions returns an igraph object built from the result
returned by \code{\link{miic}}.
}
\details{
Edges attributes are passed to the igraph graph and can be accessed with
e.g. \code{E(g)$partial_correlation}. See \code{\link{miic}} for more
details on edge parameters. By default, edges are colored according to the
partial correlation between two nodes conditioned on the conditioning set
(negative is blue, null is gray and positive is red) and their width is
based on the conditional mutual information minus the complexity cost.
}
\examples{
\dontrun{
library(miic)
library(igraph)

g = getIgraph(miic.res)
plot(g) # Default visualisation, calls igraph::plot.igraph()

# Specifying layout (see ?igraph::layout_)
l <-layout_with_kk(g)
plot(g, layout=l)

# Override some graphical parameters
plot(g, edge.curved = .2)
plot(g, vertex.shape="none", edge.color="gray85", vertex.label.color="gray10")
}

}
\seealso{
\code{\link{miic}} for details on edge parameters in the returned object,
\code{\link[igraph]{igraph.plotting}} for the detailed description of the
plotting parameters and \code{\link[igraph]{layout}} for different layouts.
}
