% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miic.export.R, R/miic.plot.R
\name{getIgraph}
\alias{getIgraph}
\title{Igraph plotting function for miic}
\usage{
getIgraph(miic.res, pcor_palette = NULL)

getIgraph(miic.res, pcor_palette = NULL)
}
\arguments{
\item{miic.res}{[a miic graph object]
The graph object returned by the miic execution.}

\item{pcor_palette}{The color palette used to represent the partial correlations
(the color of the edges). The palette must be able to handle 201 shades
to cover the correlation range from -100 to +100. The default palette is
grDevices::colorRampPalette(c("blue", "darkgrey", "red").}
}
\value{
An igraph graph object.

An igraph graph object.
}
\description{
This functions returns an igraph object built from the result
returned by \code{\link{miic}}.

This functions returns an igraph object built from the result
returned by \code{\link{miic}}.
}
\details{
Edges attributes are passed to the igraph graph and can be accessed with
e.g. \code{E(g)$partial_correlation}. See \code{\link{miic}} for more
details on edge parameters. By default, edges are colored according to the
partial correlation between two nodes conditioned on the conditioning set
(negative is blue, null is gray and positive is red) and their width is
based on the conditional mutual information minus the complexity cost.

Edges attributes are passed to the igraph graph and can be accessed with
e.g. \code{E(g)$partial_correlation}. See \code{\link{miic}} for more
details on edge parameters. By default, edges are colored according to the
partial correlation between two nodes conditioned on the conditioning set
(negative is blue, null is gray and positive is red) and their width is
based on the conditional mutual information minus the complexity cost.
}
\seealso{
\code{\link{miic}} for details on edge parameters in the returned object,
\code{\link[igraph]{igraph.plotting}} for the detailed description of the
plotting parameters and \code{\link[igraph]{layout}} for different layouts.

\code{\link{miic}} for details on edge parameters in the returned object,
\code{\link[igraph]{igraph.plotting}} for the detailed description of the
plotting parameters and \code{\link[igraph]{layout}} for different layouts.
}
