% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miic.plot.R
\name{plot.miic}
\alias{plot.miic}
\title{Basic plot function of a miic network inference result}
\usage{
\method{plot}{miic}(x, method = "igraph", pcor_palette = NULL, ...)
}
\arguments{
\item{x}{[a miic graph object]
The graph object returned by \code{\link{miic}}.}

\item{method}{A string representing the plotting method. Default to "igraph".
Currently only "igraph" is supported.}

\item{pcor_palette}{Optional. The color palette used to represent the partial
correlations (the color of the edges). See \code{\link{miic.export}} for details.}

\item{\dots}{Additional plotting parameters. See the corresponding plot function
for the complete list.
For igraph, see \code{\link[igraph]{igraph.plotting}}.}
}
\description{
This function calls \code{\link{miic.export}} to build a
plottable object from the result returned by \code{\link{miic}} and plot it.
}
\details{
See the documentation of \code{\link{miic.export}} for further
details.
}
\seealso{
\code{\link{miic.export}} for generic exports,
\code{\link{getIgraph}} for igraph export,
\code{\link[igraph]{igraph.plotting}}
}
