% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miic.R
\name{plot.miic}
\alias{plot.miic}
\title{Basic plot function of a miic network inference result}
\usage{
\method{plot}{miic}(x, ...)
}
\arguments{
\item{x}{[a miic graph object]
The graph object returned by the miic execution.}

\item{\dots}{Additional plotting parameters. See 
\code{\link[igraph]{igraph.plotting}} for the complete list.}
}
\description{
This function calls the generic \code{\link{miic.export}}
function to build an igraph object from the resuts of a 
\code{\link{miic}} call and plot it.
}
\details{
See the documentation of \code{\link{miic.export}} and 
\code{\link{getIgraph}} for further details. Requires igraph.
}
\seealso{
\code{\link{miic.export}} for generic exports,
\code{\link{getIgraph}} for igraph export,
\code{\link[igraph]{igraph.plotting}}
}
