% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectFeatures.R
\name{selectFeaturesPath}
\alias{selectFeaturesPath}
\title{Features selection on the path between variables of interest}
\usage{
selectFeaturesPath(
  input_data,
  var_of_interest_names_side1 = NULL,
  var_of_interest_values_side1 = NULL,
  var_of_interest_names_side2 = NULL,
  var_of_interest_values_side2 = NULL,
  method = "score",
  n_selected = 10,
  corrected = T,
  precomputed_mis = NULL,
  skip_cheks = F,
  n_threads = 1,
  depth_max = 10,
  verbose = 3,
  plot = F,
  ...
)
}
\arguments{
\item{input_data}{[a data frame or a matrix, required]

Expected layout is samples as rows and variables as columns. Column names
must contain the names of the variables.}

\item{var_of_interest_names_side1}{[a string or vector of strings, optional,
NULL by default]

For the variable(s) of interest that are part of the \emph{input_data}
on the first side of the path, you should supply their names here.}

\item{var_of_interest_values_side1}{[a data frame, optional, NULL by default]

For the variables of interest that are not in \emph{input_data}
and on the first side of the path, a data frame can be supplied.
The column names are the names of the variables of interest
and rows are the samples ordered in the same way as the \emph{input_data}.
Typically, such variables are metadata associated to samples but not
stored in \emph{input_data}, e.g. a "Treatment" vs "Control" variable
in an experiment and a count matrix with the expression of genes
in \emph{input_data}.}

\item{var_of_interest_names_side2}{[a string or vector of strings, optional,
NULL by default]

Same as \emph{var_of_interest_names_side1} for the second side of the path.}

\item{var_of_interest_values_side2}{[a data frame, optional, NULL by default]

Same as \emph{var_of_interest_values_side1} for the second side of the path.}

\item{method}{[a string, optional, "score" by default,  possible values: "score", "dpi"]

When set to "score", the variables selection is performed using a score
combining DPI and 3 points information (see Affeldt 2015).
By turning it to "dpi", the selection is based only on the DPI,
which speeds up the process but is less discriminating.}

\item{n_selected}{[a positive integer, optional, 10 by default]

The number of features selected at each step of the recursion.
Decreasing this value speeds up the process while reducing the number
of features selected. Increasing it has opposite effect, more features
at the cost of an increased processing time.}

\item{corrected}{[a boolean, optional, TRUE by default]

When set to TRUE, the mutual information and 3 points information
are corrected by subtracting a complexity term
(computed with the Normalized Maximum Likelihood).
For dataset having very few samples, the complexity term can have
a disproportionate impact. Setting \emph{corrected} to FALSE switches
to the use of non corrected mutual information.}

\item{precomputed_mis}{[a matrix, optional, NULL by default]

if MIs have been previously computed between some variables
in the \emph{input_data} and variable(s) of interest,
supplying these precomputed MIs speeds up the process as the existing MIs
(values present and different from NA) are not recomputed.
This matrix must have variables names from the \emph{input_data}
as row names and variables of interest names as column names
(the layout is the same as the \emph{mis} matrix returned).
To be valid, the pre-computed MI values must have been computed using
the same \emph{corrected} parameter.}

\item{skip_cheks}{[a boolean, optional, FALSE by default]

Before computing MI between the variable of interest and the features,
\emph{input_data} is checked to filter out constant features and rows full
of NAs. When the \emph{input_data} does not need such filtering,
these checks can be skipped to speed up the process.}

\item{n_threads}{[a positive integer, optional, 1 by default]

When set greater than 1, \emph{n_threads} parallel threads are used for
computation. Make sure your compiler is compatible with openmp
if you wish to use multithreading.}

\item{depth_max}{[a positive integer, optional, 10 by default]

The maximum depth of the recursion.}

\item{verbose}{[an integer, optional, 3 by default]

Level of verbosity: 0=no display, 1=summary, 2=progress per couple of
variable of interest, 3=include more detail on the progress during
the recursion.}

\item{plot}{[a boolean, optional, FALSE by default]

If set to TRUE, a plot with the top features is generated.

Please note on the plot rendering that positions are indicative with
a tendancy to be displayed to the left side: a feature can appear multiple
times during the recursion, but couples already investigated are skipped
to speed up the process. For the plotting, the rule applied to select the
position of each feature is (from the positions returned by the recursion),
to pick the one corresponding to the minimal depth.}

\item{...}{If plotting is requested, extra parameters can be used to customize the plot
rendering:

\itemize{
\item \emph{x_lab}: a string, optional, "Features for
  variables_of_interest_side1-variables_of_interest_side2" by default,
  the X axis label.
\item \emph{depth_plot}: an integer, optional, 4 by default,
  maximal depth used for the plot.
\item \emph{annotate}: a boolean, optional, TRUE by default,
  When activated, displays the maximal depth of the recursion
  and the number of features not displayed on the plot.
\item \emph{font_size}: an integer, optional, 11 by default, the font size.
}}
}
\value{
A named list with five items:

\itemize{
\item \emph{features}: a vector with the features selected.
\item \emph{mis}: a matrix with the MIs between the variables in
  \emph{input_data} (as rows) and the variable(s) of interest (as columns).
  Row and column names are sorted alphabetically.\cr
  The MIs can be corrected or not, depending on the \emph{corrected}
  parameters.
  If a pre-computed MIs matrix was supplied, new values computed
  are added to the existing matrix.
\item \emph{couples}: a data frames with, at each depth, the couples
  of variables used with their MIs and features selected.
  Please note that couples already encountered are not included
  as no recursion is performed on duplicates.
\item \emph{scores}: a data frame with the information about all pairs of
  variables passing the dpi test, containing the depth, the MI values,
  the DPI test and, if \emph{method} is "score", the 3 points information
  and score.
\item \emph{plot}: when the \emph{plot} parameter is turned to TRUE,
  a plot with the features of the top levels of recursion, NULL otherwise.
}
}
\description{
Select the variables that are the most likely to be in the path
between two set of variable(s) of interest.
Variables selection can be performed using a score combining Data Processing
inequality (DPI) and 3 points information or only using the DPI.\cr
The search is recursive by dichotomy: in the first round, a set of features
on the path between the main variables of interest is selected.
From then, the path is splitted in two parts:
\itemize{
  \item variables of interest of side 1 + set of features selected
  \item set of features selected + variables of interest of side 2
  }
Each part of the path is investigated with the same principle:
finding a set of features in the path subsection, splitting the path in two,
and so on.
The recursion ends when no feature can be found
or when the maximum level of recursion is reached.
}
\examples{
library(miic)

# Features selection on the path between an external metadata "Ploidy"
# (simulation by extracting "Ploidy" out of the dataset)
# and the gene expression of TP53
df_external_meta <- data.frame ("Ploidy" = cosmicCancer$Ploidy)
df_data <- cosmicCancer[ , ! (colnames(cosmicCancer) == "Ploidy") ]
ret <- selectFeaturesPath (df_data,
                           var_of_interest_values_side1=df_external_meta,
                           var_of_interest_names_side2="TP53")
message ("Features selected = ", paste (ret$features, collapse=", ") )

# Same features selection with reuse of the MIs computed above and plot
if ( require(ggplot2) ) {
  ret <- selectFeaturesPath (df_data,
                             var_of_interest_values_side1=df_external_meta,
                             var_of_interest_names_side2="TP53",
                             precomputed_mis=ret$mis,
                             plot=TRUE)
  print (ret$plot)
}

# Features selection using multiple variables on each side
ret <- selectFeaturesPath (cosmicCancer,
  var_of_interest_names_side1=c("TP53", "Ploidy"),
  var_of_interest_names_side2=c("FOXM1", "AURKA") )
message ("Features selected = ", paste (ret$features, collapse=", ") )

# Same features selection with plotting using a customized rendering
if ( require(ggplot2) ) {
  ret <- selectFeaturesPath (cosmicCancer,
                            var_of_interest_names_side1=c("TP53", "Ploidy"),
                            var_of_interest_names_side2=c("FOXM1", "AURKA"),
                            plot=TRUE, depth_plot=6, font_size=9)
  print (ret$plot)
}

}
\references{
\itemize{
\item Affeldt \emph{et al.}, UAI 2015, \href{https://auai.org/uai2015/proceedings/papers/293.pdf}{Robust Reconstruction of Causal Graphical Models based on Conditional 2-point and 3-point Information}
}
}
