% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectFeatures.R
\name{selectFeaturesPath}
\alias{selectFeaturesPath}
\title{Features selection on the path between two sets of variables of interest
based on Mutual Information (MI)}
\usage{
selectFeaturesPath(
  input_data,
  n_features,
  var_of_interest_names_side1 = NULL,
  var_of_interest_values_side1 = NULL,
  var_of_interest_names_side2 = NULL,
  var_of_interest_values_side2 = NULL,
  skip_cheks = F,
  precomputed_mis = NULL,
  filter_v_struct = F,
  n_threads = 1,
  verbose = 3,
  plot = F,
  ...
)
}
\arguments{
\item{input_data}{[a data frame or a matrix, required]

Expected layout is samples as rows and variables as columns. Column names
must contain the names of the variables.}

\item{n_features}{[an integer, required]

The total number of features to select, must be in the range [0, number
of variables in \emph{input_data}]. If several variables of interest
are supplied, the function will try to select an equal number of features
per variable.

When \emph{n_features} is 0, no feature selection will be performed
but the MIs between the variables in \emph{input_data} and the variable(s)
of interest will be computed and returned.
This can be used to compute once for all the MIs and perform several
features selection at a later stage, supplying MIs matrix in the
\emph{precomputed_mis} parameter.}

\item{var_of_interest_names_side1}{[a string or vector of strings, optional,
NULL by default]

For the variable(s) of interest that are part of the \emph{input_data}
on the first side of the path, you should supply their names here.}

\item{var_of_interest_values_side1}{[a data frame, optional, NULL by default]

For the variables of interest that are not in \emph{input_data}
and on the first side of the path, a data frame can be supplied.
The column names are the names of the variables of interest
and rows are the samples ordered in the same way as the \emph{input_data}.
Typically, such variables are metadata associated to samples but not
stored in \emph{input_data}, e.g. a "Treatment" vs "Control" variable
in an experiment and a count matrix with the expression of genes
in \emph{input_data}.}

\item{var_of_interest_names_side2}{[a string or vector of strings, optional,
NULL by default]

Same as \emph{var_of_interest_names_side1} for the second side of the path.}

\item{var_of_interest_values_side2}{[a data frame, optional, NULL by default]

Same as \emph{var_of_interest_values_side1} for the second side of the path.}

\item{skip_cheks}{[a boolean, optional, FALSE by default]

Before computing MI between the variable of interest and the features,
\emph{input_data} is checked to filter out constant features and rows full
of NAs. When the \emph{input_data} does not need such filtering,
these checks can be skipped to speed up the process.}

\item{precomputed_mis}{[a matrix, optional, NULL by default]

if MIs has been previously computed between some variables
in the \emph{input_data} and variable(s) of interest,
supplying these precomputed MIs will speed up the process as the existing
MIs (values different from NA) will not be recomputed.
This matrix must have variables names from the \emph{input_data}
as row names and variables of interest names as column names
(the layout is the same as the \emph{mis} matrix returned).}

\item{filter_v_struct}{[a boolean, optional, FALSE by default]

If set to TRUE, the features passing the DPI test are filtered
out if they form a v-structure with the couple of variables of interest.}

\item{n_threads}{[a positive integer, optional, 1 by default]

When set greater than 1, \emph{n_threads} parallel threads will be used for
computation. Make sure your compiler is compatible with openmp
if you wish to use multithreading.}

\item{verbose}{[an integer, optional, 3 by default]

Level of verbosity: 0=no display, 1=summary, 2=progress per variable of
interest, 3=same as 2 with display of estimated time remaining.}

\item{plot}{[a boolean, optional, FALSE by default]

If set to TRUE, a plot with the top features for each couple of variables of
interest will be generated.

Extra parameters can be used to customize the plot rendering:

\itemize{
\item \emph{n_plots}: an integer, optional, 25 by default, the number of
  features to plot.
\item \emph{x_lab}: a string, optional, "Top features for
  variable_of_interest_1-variable_of_interest_2" by default,
  the X axis label.
\item \emph{y_lab}: a string, optional, "MI (bits)" by default,
  the Y axis label.
\item \emph{values}: a boolean, optional, TRUE by default,
  displays the MI values of each feature when activated.
\item \emph{annotate}: a boolean, optional, TRUE by default,
  displays the total number of features passing the DPI test
  when activated.
\item \emph{threshold}: a boolean, optional, TRUE by default,
  displays the MI threshold of the DPI test when activated.
\item \emph{font_size}: an integer, optional, 11 by default, the font size.
\item \emph{box_fill}: a vector, optional, c("#A6CEE3", "#1F78B4") by
  default (paired blues of brewer palette), the bar plot filling colors.
\item \emph{box_text}: a string, optional, "white" by default,
  the text color of MI values in the bar plot.
}}
}
\value{
A named list with six items:

\itemize{
\item \emph{features}: a vector with the features selected.
\item \emph{mis}: a matrix with the MIs (in bits) between the variables in
  \emph{input_data} (as rows) and the variable(s) of interest (as columns).
  Row and column names are sorted alphabetically.\cr
  If a precomputed MIs matrix was supplied, new values computed
  are added to the existing matrix.
\item \emph{couples}: a data frames with the MIs of each couple of variables
  of interest used as threshold for the DPI test.
\item \emph{dpis}: a list if data frames with the MIs (in bits) used for the
  DPI tests. Each item corresponds to a couple of variables of interest.
  The row names of the data frames are the features passing the DPI test
  and data frames are ordered by sum of MIs with the variables of interest
  descending.
\item \emph{plots}: when the \emph{plot} parameter is turned to TRUE,
  a list of plots with the top features for each couple of variables of
  interest. The list is empty if the \emph{plot} parameter is set to FALSE.
\item \emph{ais}: when the \emph{filter_v_struct} parameter is turned to
  TRUE, a list with the separating set for each couple of variables of
  interest.
  The list is empty if the \emph{filter_v_struct} parameter is set to FALSE.
\item \emph{v_struct_excluded}: the \emph{filter_v_struct} parameter is
  turned to TRUE, a list with the features excluded (because forming a
  v-structure) for each couple of variables of interest.
  The list is empty if the \emph{filter_v_struct} parameter is set to FALSE.
}
}
\description{
Select features by using data propagation inequality (DPI)
to select the set of variables that are the most likely to be in the path
between two set of variable(s) of interest.
When multiple variables of interest are supplied in one or both sides of
the path, the function will try to select an equal number of features
per couple (one of each side) of variables of interest .

Variables not sharing information at all (MI = 0) or not in the path
between the set of variables considering the DPI can not be selected
as features. So the number of features returned globally or linked to a
set of variables of interest can be lower than requested. This can also
lead to unbalanced number of features per couple of variables of interest.
}
\examples{
library(miic)

# Features selection on the path between an external metadata "Ploidy"
# (simulation by extracting "Ploidy" out of the dataset)
# and the gene expression of TP53
df_external_meta <- data.frame ("Ploidy" = cosmicCancer$Ploidy)
df_data <- cosmicCancer[ , ! (colnames(cosmicCancer) == "Ploidy") ]
ret <- selectFeaturesPath (df_data,
                           var_of_interest_values_side1=df_external_meta,
                           var_of_interest_names_side2="TP53",
                           n_features=10)
message ("Tops 10 features = ", paste (ret$features, collapse=", ") )

# Same features selection with reuse of the MIs computed above and plot
if ( require(ggplot2) ) {
  ret <- selectFeaturesPath (df_data,
                             var_of_interest_values_side1=df_external_meta,
                             var_of_interest_names_side2="TP53",
                             precomputed_mis=ret$mis,
                             n_features=10,
                             plot=TRUE)
  print (ret$plots[[1]])
}

# Features selection using multiple variables on each side
ret <- selectFeaturesPath (cosmicCancer,
  var_of_interest_names_side1=c("TP53", "Ploidy"),
  var_of_interest_names_side2=c("FOXM1", "AURKA"),
  n_features=10)
message ("Tops 10 features = ", paste (ret$features, collapse=", ") )

# Same features selection with plotting using a customized rendering
if ( require(ggplot2) && require(gridExtra) ) {
  ret <- selectFeaturesPath (cosmicCancer,
                             var_of_interest_names_side1=c("TP53", "Ploidy"),
                             var_of_interest_names_side2=c("FOXM1", "AURKA"),
                             n_features=10,
                             plot=TRUE,
                             n_plots=10,
                             box_fill=c("lawngreen", "limegreen"),
                             box_text="black")
  do.call ("grid.arrange", c(ret$plots, ncol=2,
    top="Top features on cosmicCancer dataset") )
}

}
