% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmiic.wrapper.R
\name{tmiic.flatten_network}
\alias{tmiic.flatten_network}
\title{tmiic.flatten_network}
\usage{
tmiic.flatten_network(
  tmiic.res,
  flatten_mode = "compact",
  keep_edges_on_same_node = TRUE
)
}
\arguments{
\item{tmiic.res}{[a tmiic object] The object returned by miic's 
execution in temporal mode.}

\item{flatten_mode}{[a string]. Optional, default value "compact".
Possible values are \emph{"compact"}, \emph{"combine"}, \emph{"unique"}, 
\emph{"drop"}:
\itemize{
\item When \emph{flatten_mode} = \emph{"compact"}, the default. Nodes 
  and edges are converted into a flattened version preserving all 
  the initial information.\cr
  i.e.: X_lag1->Y_lag0, X_lag0<-Y_lag2 become respectively X->Y lag=1, 
  X<-Y lag=2. 
\item When \emph{flatten_mode} = \emph{"combine"}, one edge will be kept
  per couple of nodes. The info_shifted will be the highest of the 
  summarized edges whilst the lag and orientation of the summarized 
  edge will be an agregation.\cr 
  i.e.: X_lag2->Y_lag0, X_lag0<-Y_lag1 will become X<->Y lag=1,2 with
  the info_shifted of X_lag2->Y_lag0 if info_shifted of 
  X_lag2->Y_lag0 > X_lag0<-Y_lag1.
\item When \emph{flatten_mode} = \emph{"unique"}, only the edges having the
  highest info_shifted for a couple of nodes are kept in the flattened 
  network. If several edges between the sames nodes have the same
  info_shifted, then the edge kept is the one with the minimum lag.\cr 
  i.e.: X_lag1->Y_lag0, X_lag0<-Y_lag2 with info_shifted of 
  X_lag1->Y_lag0 > X_lag0<-Y_lag2 become X->Y lag=1.
\item When \emph{flatten_mode} = \emph{"drop"}, only the edges having the 
  highest info_shifted for a couple of nodes are kept in the flattened 
  network. If several edges between the sames nodes have the same
  info_shifted, then the edge kept is the one with the minimum lag.\cr
  i.e. :  X_lag1->Y_lag0, X_lag0<-Y_lag2 with info_shifted of 
  X_lag1->Y_lag0 > X_lag0<-Y_lag2 become X->Y. The lag information is 
  "lost" after flattening
}
Note that for all modes other than \emph{"drop"}, lag is a new column 
added in the dataframe.}

\item{keep_edges_on_same_node}{[a boolean] Optional, TRUE by default.
When TRUE, the edges like X_lag0-X_lag1 are kept during flattening
(it becomes an X-X edge). When FALSE, only edges having different nodes 
are kept in the flatten network.}
}
\value{
[a tmiic object] The returned tmiic object is the one received 
as input where the summary dataframe has been flattened.
}
\description{
Flattten the lagged network returned by tmiic
}
\details{
In temporal mode, the network returned by miic contains lagged nodes 
(X_lag0, X_lag1, ...). This function flatten the  network depending 
of the \emph{flatten_mode} parameter.\cr
Note that only the summary data frame is flattened.
}
