% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmiic.plot.R
\name{tmiic.getIgraph}
\alias{tmiic.getIgraph}
\title{Igraph plotting function for tmiic (temporal mode of miic)}
\usage{
tmiic.getIgraph(tmiic.res)
}
\arguments{
\item{tmiic.res}{[a tmiic graph object]
The graph object returned by the miic execution in temporal mode, 
eventually flattened}
}
\value{
An igraph graph object.
}
\description{
This functions returns an igraph object built from the result
returned by:
\code{\link{miic}} executed in temporal temporal mode.
\code{\link{tmiic.flatten_network}}
}
\details{
Edges attributes are passed to the igraph graph and can be accessed with
e.g. \code{E(g)$partial_correlation}. See \code{\link{miic}} for more
details on edge parameters. By default, edges are colored according to the
partial correlation between two nodes conditioned on the conditioning set
(negative is blue, null is gray and positive is red) and their width is
based on the conditional mutual information minus the complexity cost.
}
\seealso{
\code{\link{miic}} for details on edge parameters in the returned object,
\code{\link[igraph]{igraph.plotting}} for the detailed description of the
plotting parameters and \code{\link[igraph]{layout}} for different layouts.
}
